end_to_end tests use client
